Add files via upload
Switched the dynamics of the website to ensure that it is all on one page. You can manually scroll through all pages and I also implemented the tabs in the header to automatically scroll to the given page/section when clicked. Also added a pop up page when clicking on "book" from the pricing section. This pop up page will be the form. Finally, I added images and testimonials to the repairs page.